pyproject.toml
[project]
name = "jingshan-toolkit"
version = "0.1.0"
description = "京城山海经·新编 —— RunningHub AIGC 调用 / 归档工具"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "typer>=0.12",
    "rich>=13.7",
    "python-dotenv>=1.0",
    "tomli>=2.0; python_version < '3.11'",
]

[project.scripts]
rh = "jingshan.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["jingshan"]
